// frontend/src/api_client/models/ProductCategory.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum ProductCategory {
    HOUSEHOLD_PERSONAL = 'HOUSEHOLD_PERSONAL',
    VEHICLE = 'VEHICLE',
    DOCUMENTS = 'DOCUMENTS',
    HAZARDOUS = 'HAZARDOUS',
    COMMERCIAL = 'COMMERCIAL',
    ELECTRONICS = 'ELECTRONICS',
    APPAREL = 'APPAREL',
    OTHER = 'OTHER',
}